api/pyproject.toml
[tool.poetry]
name = "cloud-web-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.2.3"
flask-restful = "0.3.9"
gunicorn = "20.1.0"
pyjwt = "2.6.0"
python-dotenv = "^1.0.0"
cloud-db = {git = "https://github.com/b-munar/cloud-db.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
